+++ extensions/ima/src/main/java/com/google/android/exoplayer2/ext/ima/ImaAdsLoader.java
@@ -58,6 +58,7 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.Locale;
 import java.util.Set;
 
 /**
@@ -544,7 +545,9 @@ public final class ImaAdsLoader implements Player.EventListener, AdsLoader {
   private static final class DefaultImaFactory implements ImaUtil.ImaFactory {
     @Override
     public ImaSdkSettings createImaSdkSettings() {
-      return ImaSdkFactory.getInstance().createImaSdkSettings();
+      ImaSdkSettings settings = ImaSdkFactory.getInstance().createImaSdkSettings();
+      settings.setLanguage(getImaLanguageCodeForDefaultLocale());
+      return settings;
     }
 
     @Override
@@ -584,6 +587,17 @@ public final class ImaAdsLoader implements Player.EventListener, AdsLoader {
         Context context, ImaSdkSettings imaSdkSettings, AdDisplayContainer adDisplayContainer) {
       return ImaSdkFactory.getInstance()
           .createAdsLoader(context, imaSdkSettings, adDisplayContainer);
+    }
+
+    /**
+     * Returns a language code that's suitable for passing to {@link ImaSdkSettings#setLanguage} and
+     * corresponds to the device's {@link Locale#getDefault() default Locale}.
+     */
+    // TODO: It may be possible to define a better mapping onto IMA's supported language codes. See:
+    // https://developers.google.com/interactive-media-ads/docs/sdks/android/client-side/localization.
+    // IMA will fall back to its default language code ("en") if the value returned is unsupported.
+    private static String getImaLanguageCodeForDefaultLocale() {
+      return Util.splitAtFirst(Util.getSystemLanguageCodes()[0], "-")[0];
     }
   }
 }